refactor(shifts): シフト管理画面を「シフト変更履歴」専用画面に再編
業務管理画面 (/duty-assignments) がシフト編集の主画面として
確立されたため、シフト管理画面の月次/日次カレンダーおよび編集
タブを廃止。サイドバーラベルを「シフト変更履歴」に改称し、URL
を /shifts/history に統一。旧 /shifts および ?tab=history は
互換リダイレクトで維持。

- 新 /shifts/history/page.tsx を実装ページに格上げ
- /shifts/page.tsx を /shifts/history への互換リダイレクトに
- 履歴詳細画面のパンくずを「ダッシュボード > シフト変更履歴 > 履歴 #N」に
- shift-history-detail.tsx の戻り先 URL 3 箇所を整理
- サイドバー: ラベル/href/アイコンを更新 (Calendar → History)
- shift-actions.ts: revalidatePath を /shifts/history と /duty-assignments
  に統一。bulkUpdateShifts/importShifts には /duty-assignments を補修
- lib/constants.ts の未参照 NAV_ITEMS をブロック削除
- 専用コンポーネント 8 ファイルと loading.tsx を削除
- E2E (sidebar.spec.ts) を新ラベル・新 URL に更新
