--- components/shifts/shift-calendar-cell.tsx
@@ -1,89 +1,0 @@
-"use client"
-
-import type { Shift } from "@/app/generated/prisma/client"
-import { getShiftCodeInfo, type ShiftCodeInfo } from "@/lib/constants"
-import { formatTime } from "@/lib/date-utils"
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipTrigger,
-} from "@/components/ui/tooltip"
-import { cn } from "@/lib/utils"
-import { Home } from "lucide-react"
-
-type ShiftCalendarCellProps = {
-  shift?: Shift
-  isWeekend: boolean
-  isToday: boolean
-  isSelected: boolean
-  hasHistory?: boolean
-  onClick: () => void
-  onSelect?: () => void
-  shiftCodeMap?: Record<string, ShiftCodeInfo>
-}
-
-export function ShiftCalendarCell({
-  shift,
-  isWeekend,
-  isToday,
-  isSelected,
-  hasHistory,
-  onClick,
-  onSelect,
-  shiftCodeMap,
-}: ShiftCalendarCellProps) {
-  const code = shift?.shiftCode ?? null
-  const info = getShiftCodeInfo(code, shiftCodeMap)
-
-  const handleClick = (e: React.MouseEvent) => {
-    if (e.shiftKey && onSelect) {
-      e.preventDefault()
-      onSelect()
-    } else {
-      onClick()
-    }
-  }
-
-  return (
-    <Tooltip>
-      <TooltipTrigger asChild>
-        <button
-          type="button"
-          onClick={handleClick}
-          className={cn(
-            "relative h-full w-full text-xs font-medium transition-colors cursor-pointer",
-            "border-r border-b border-border/50",
-            "hover:ring-2 hover:ring-primary/30 hover:z-10",
-            shift?.isHoliday ? "text-red-600" : info.color,
-            isToday && "bg-primary/10",
-            isWeekend && "bg-red-50",
-            isToday && isWeekend && "bg-red-100",
-            hasHistory && "bg-sky-100",
-            isSelected && "ring-2 ring-primary bg-primary/10 z-20"
-          )}
-        >
-          {code ?? ""}
-          {shift?.isRemote && (
-            <Home className="absolute top-0 right-0 h-2.5 w-2.5 text-sky-600" />
-          )}
-        </button>
-      </TooltipTrigger>
-      {shift && (
-        <TooltipContent side="top" className="text-xs">
-          <div className="space-y-1">
-            <div className="font-medium">{info.label}</div>
-            {shift.startTime && (
-              <div>
-                {formatTime(shift.startTime)} - {formatTime(shift.endTime)}
-              </div>
-            )}
-            <div className="flex gap-1">
-              {shift.isHoliday && <span className="text-red-600">休日</span>}
-              {shift.isRemote && <span className="text-sky-600">テレワーク</span>}
-            </div>
-          </div>
-        </TooltipContent>
-      )}
-    </Tooltip>
-  )
-}
