+++ lib/constants.ts
@@ -63,15 +63,3 @@ export function getColorClasses(colorKey: string | null): { text: string; bg: st
 // ページサイズ
 export const PAGE_SIZES = [10, 20, 50, 100] as const
 export const DEFAULT_PAGE_SIZE = 20
-
-// ナビゲーション
-export const NAV_ITEMS = [
-  { label: "ダッシュボード", href: "/", icon: "LayoutDashboard" },
-  { label: "シフト管理", href: "/shifts", icon: "Calendar" },
-  { label: "シフト履歴", href: "/shifts/history", icon: "History" },
-  { label: "従業員", href: "/employees", icon: "Users" },
-  { label: "グループ", href: "/groups", icon: "FolderOpen" },
-  { label: "ロール", href: "/roles", icon: "Shield" },
-  { label: "役職", href: "/positions", icon: "Award" },
-  { label: "シフトコード", href: "/shift-codes", icon: "Tag" },
-] as const
